sample_code_1/requirements.txt
# E-Commerce Multi-Agent System Dependencies
# Core Framework Dependencies
networkx>=3.4.2
pyyaml>=6.0.1
asyncio>=3.4.3
rich>=14.0.0
pydantic>=2.11.3

# LLM and AI Dependencies
google-genai>=0.8.0
openai>=1.50.0
anthropic>=0.25.0
aiohttp>=3.9.0
httpx>=0.28.1

# MCP (Model Context Protocol) Dependencies
mcp[cli]>=1.7.0
mcp[stdio]>=1.7.0
mcp[sse]>=1.7.0

# Database and Storage Dependencies
psycopg2-binary>=2.9.9
redis>=5.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# File Processing Dependencies
pillow>=11.2.1
pymupdf4llm>=0.0.21
trafilatura[all]>=2.0.0
markitdown[all]>=0.1.1
pandas>=2.2.0
numpy>=1.26.0

# Web and API Dependencies
fastapi>=0.115.12
uvicorn>=0.34.2
sse-starlette>=2.3.4
requests>=2.32.3
beautifulsoup4>=4.12.0

# Utility Dependencies
python-dotenv>=1.0.0
click>=8.1.0
tqdm>=4.67.1
python-multipart>=0.0.6
jinja2>=3.1.6

# Testing Dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development Dependencies
black>=24.0.0
flake8>=7.0.0
mypy>=1.8.0
pre-commit>=3.6.0

# Monitoring and Logging
structlog>=24.1.0
prometheus-client>=0.19.0

# Security Dependencies
cryptography>=42.0.0
passlib>=1.7.4

# Optional: Local LLM Support
# ollama>=0.1.0  # Uncomment if using local Ollama models

# Optional: Vector Database Support
# faiss-cpu>=1.10.0  # Uncomment if using FAISS for embeddings
# chromadb>=0.4.0  # Uncomment if using ChromaDB

# Optional: Browser Automation
# playwright>=1.52.0  # Uncomment if using browser automation

# Optional: Advanced Analytics
# matplotlib>=3.10.3  # Uncomment if using plotting
# seaborn>=0.13.0  # Uncomment if using statistical plots 